Application/BackEnd/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
requests
psycopg2-binary
python-dotenv
google-generativeai